added some extra authentication codes

+++ src/main/java/com/pratik/JournalApp/service/JournalEntryService.java
@@ -28,11 +28,11 @@ private JournalEntryRepository journalEntryRepository;
           User user = userService.findByUserName(userName);
           JournalEntry saved =journalEntryRepository.save(journalEntry);
           user.getJournalEntries().add(saved);
-          userService.saveEntry(user);
+          userService.saveUser(user);
 
       } catch (Exception e){
           System.out.println(e);
-          throw new RuntimeException("Any error occurred while saving the entry. ",e);
+          throw new RuntimeException("Any error occurred while saving the entry.",e);
       }
 
    }
@@ -50,11 +50,28 @@ private JournalEntryRepository journalEntryRepository;
      return journalEntryRepository.findById(id);
    }
 
-   public void deleteById(ObjectId id, String userName){
-      User user = userService.findByUserName(userName);
-      user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-      userService.saveEntry(user);
-      journalEntryRepository.deleteById(id);
+   @Transactional
+   public boolean deleteById(ObjectId id, String userName){
+
+      boolean removed = false;
+      try{
+          User user = userService.findByUserName(userName);
+           removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+          if(removed){
+              userService.saveUser(user);
+              journalEntryRepository.deleteById(id);
+          }
+
+      }catch (Exception e){
+          System.out.println(e);
+          throw new RuntimeException("An error occured while deleting the entry", e);
+      }
+       return removed;
+
    }
+
+//   public List<JournalEntry> findByUserName(String userName){
+//    return journalEntryRepository.
+//   }
 }
 //Controller---->Service---->Repository
